+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release Notes for Blitz Netlify Deployer
 
+## 2.0.0-beta.1 - 2022-04-19
+### Added
+- Added compatibility with Craft 4.
+
+### Changed
+- Changed minimum required version of Blitz to 4.0.0-beta.2.
+
 ## 1.0.1 - 2019-12-18
 ### Changed
 - Increased minimum required Blitz version to 3.0.2.